chore: extract `DeviceRequestPrompt` interface
